Add outstandingUsers property to Bet
This will allow us to keep track of users who have lost bets and prompt
them to send proof to other people involved in the bet that they have
fulfilled their bet stake. Eventually we will use this to keep track
of users that need to upload videos of their shotguns.